+++ Tunny/Optimization/RestoreLoop.cs
@@ -3,12 +3,11 @@ using System.ComponentModel;
 using System.Linq;
 using System.Windows.Forms;
 
-using Rhino.FileIO;
 using Rhino.Geometry;
 
 using Tunny.Component;
-using Tunny.GHType;
 using Tunny.Solver;
+using Tunny.Type;
 using Tunny.UI;
 using Tunny.Util;
 
@@ -28,7 +27,7 @@ namespace Tunny.Optimization
             s_worker = sender as BackgroundWorker;
             s_component = e.Argument as TunnyComponent;
 
-            var cFishes = new List<Fish>();
+            var fishes = new List<Fish>();
 
             var optunaSolver = new Optuna(s_component.GhInOut.ComponentFolder);
             ModelResult[] modelResult = optunaSolver.GetModelResult(Indices, StudyName);
@@ -43,7 +42,7 @@ namespace Tunny.Optimization
                 case "Restore":
                     for (int i = 0; i < modelResult.Length; i++)
                     {
-                        SetResultToFish(cFishes, modelResult[i], NickNames);
+                        SetResultToFish(fishes, modelResult[i], NickNames);
                         s_worker.ReportProgress(i * 100 / modelResult.Length);
                     }
                     break;
@@ -55,12 +54,12 @@ namespace Tunny.Optimization
                             "Tunny"
                         );
                     }
-                    SetResultToFish(cFishes, modelResult[0], NickNames);
+                    SetResultToFish(fishes, modelResult[0], NickNames);
                     s_worker.ReportProgress(100);
                     break;
             }
 
-            s_component.Fishes = cFishes.ToArray();
+            s_component.Fishes = fishes.ToArray();
             s_worker.ReportProgress(100);
 
             if (s_worker != null)
@@ -77,7 +76,7 @@ namespace Tunny.Optimization
                 ModelNumber = model.Number,
                 Variables = SetVariables(model, nickname),
                 Objectives = SetObjectives(model),
-                Geometries = SetGeometries(model)
+                Attributes = SetAttributes(model),
             });
         }
 
@@ -98,6 +97,10 @@ namespace Tunny.Optimization
         {
             string[] nickNames = s_component.GhInOut.Objectives.Select(x => x.NickName).ToArray();
             var objectives = new Dictionary<string, double>();
+            if (model.Objectives == null)
+            {
+                return null;
+            }
             for (int i = 0; i < model.Objectives.Length; i++)
             {
                 objectives.Add(nickNames[i], model.Objectives[i]);
@@ -105,20 +108,26 @@ namespace Tunny.Optimization
             return objectives;
         }
 
-        private static List<GeometryBase> SetGeometries(ModelResult model)
+        private static Dictionary<string, object> SetAttributes(ModelResult model)
         {
-            var geometries = new List<GeometryBase>();
-            if (model.GeometryJson.Length == 0)
+            var attribute = new Dictionary<string, object>();
+            foreach (KeyValuePair<string, List<string>> attr in model.Attributes)
             {
-                return geometries;
+                if (attr.Key == "Geometry")
+                {
+                    var geometries = new List<GeometryBase>();
+                    foreach (string json in attr.Value)
+                    {
+                        geometries.Add(Rhino.Runtime.CommonObject.FromJSON(json) as GeometryBase);
+                    }
+                    attribute.Add(attr.Key, geometries);
+                }
+                else
+                {
+                    attribute.Add(attr.Key, attr.Value);
+                }
             }
-
-            foreach (string json in model.GeometryJson)
-            {
-                var geometry = Rhino.Runtime.CommonObject.FromJSON(json) as GeometryBase;
-                geometries.Add(geometry);
-            }
-            return geometries;
+            return attribute;
         }
     }
 }