+++ Tunny/Resources/Resource.Designer.cs
@@ -63,9 +63,49 @@ namespace Tunny.Resources {
         /// <summary>
         ///   型 System.Drawing.Bitmap のローカライズされたリソースを検索します。
         /// </summary>
+        internal static System.Drawing.Bitmap ConstructFishAttribute {
+            get {
+                object obj = ResourceManager.GetObject("ConstructFishAttribute", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   型 System.Drawing.Bitmap のローカライズされたリソースを検索します。
+        /// </summary>
+        internal static System.Drawing.Bitmap DeconstructFish {
+            get {
+                object obj = ResourceManager.GetObject("DeconstructFish", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   型 System.Drawing.Bitmap のローカライズされたリソースを検索します。
+        /// </summary>
+        internal static System.Drawing.Bitmap DeconstructFishAttribute {
+            get {
+                object obj = ResourceManager.GetObject("DeconstructFishAttribute", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   型 System.Drawing.Bitmap のローカライズされたリソースを検索します。
+        /// </summary>
         internal static System.Drawing.Bitmap FishMarket {
             get {
                 object obj = ResourceManager.GetObject("FishMarket", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   型 System.Drawing.Bitmap のローカライズされたリソースを検索します。
+        /// </summary>
+        internal static System.Drawing.Bitmap ParamFishAttribute {
+            get {
+                object obj = ResourceManager.GetObject("ParamFishAttribute", resourceCulture);
                 return ((System.Drawing.Bitmap)(obj));
             }
         }